akira-web/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
boto3==1.42.75
botocore==1.42.75
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comtypes==1.4.16
fastapi==0.135.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
lxml==6.0.2
MouseInfo==0.1.3
mss==10.1.0
numpy==2.4.3
opencv-python==4.13.0.92
pillow==12.1.1
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytweening==1.2.0
pywin32==311
pywinauto==0.6.9
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
starlette==1.0.0
tinycss2==1.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
litellm>=1.55.0,<2
PyYAML>=6.0.1
webencodings==0.5.1

torch==2.3.1
# flash_attn==2.5.8  # Commented out: Linux-only, requires CUDA build
accelerate==0.31.0
transformers==4.41.2
huggingface-hub>=0.23.0,<1.0